requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
idna==2.10
Pillow==8.0.1
pycodestyle==2.6.0
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
